Allow running acceptance tests with system `terraform`
Terraform pre-releases are not built to Dockerhub. Allow passing path to
a `terraform` executable to run the acceptance tests with it. The
version must of course match to the specified one.

We can't simply `exec terraform` as the exact same wrapper script is
first in the PATH, thus leading to infinite recursion.

Example:

    $ chtf 0.12.0-alpha4
    $ USE_SYSTEM_TERRAFORM=$(which terraform) \
        bundle exec rake 'test:acceptance[0.12.0-alpha4]'