Copy the lock file after initializing Terraform
Terraform does not follow the symlink, but replaces it with new file.
So copy the generated/updated file back to the module dir right after
calling `terraform init`.

This seems safe and consistent also in case of (Terraform) failures.
It also feels best to use only one lock file for all environments.

Downside of the YleTf's init behaviour is that the Terraform output
is "hidden" as debug output, and that the `terraform init` command
is hardcoded. Add two features to ease up a bit:

- Support passing arguments to the automatic init call with
  `TF_INIT_ARGS` environment variable.

- Skip the internal init call if `tf <env> init [args]` is called. This
  makes it possible to e.g. `-upgrade` the dependencies.