Support encrypting file backend states
Add support for automatically encrypting and decrypting the state file
when using the (default) `file` backend. The feature is still
experimental as functionality might change after gaining experience and
feedback.

If encryption is enabled, instead of symlinking, the state file is
decrypted from the `path` to the temporary working directory just before
initializing Terraform, and decrypted back after running the actual
command.

If the encryption fails, the uncrypted state file is copied to
"`path`.plaintext" for the user to encrypt manually.

By default the encryption is done using [sops](https://github.com/mozilla/sops#readme),
but the whole file is encrypted instead of just leaf values, as the
whole content is anyway encrypted on every run.
Sops can be configured further (e.g. the keys) with `.sops.yaml`.

The actual commands are also configurable, for example:

backend:
  file:
    encrypt: true
    encrypt_command: sh -c 'tr "A-Za-z" "N-ZA-Mn-za-m" <"{{FROM}}" >"{{TO}}"'
    decrypt_command: sh -c 'tr "A-Za-z" "N-ZA-Mn-za-m" <"{{FROM}}" >"{{TO}}"'