[tf_hook_output_logger] Use force_encoding instead of encoding. (#44)
When we get output from terraform hook, the problem is that content is labeled as US-ASCII and String.encode won't solve that one. However, String.force_encoding will let ruby know that content is in UTF-8, even if it has originally assumed that it is something ele.
Rescue tries to guarantee that our deployment won't suffer complete failure in post-hook phase.